social-media-scraper/requirements.txt
asttokens==2.4.0
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.1.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
comm==0.1.4
cryptography==41.0.4
debugpy==1.8.0
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
jupyter_client==8.3.1
jupyter_core==5.3.2
kaitaistruct==0.10
langdetect==1.0.9
lxml==4.9.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
nested-lookup==0.2.25
numpy==1.26.0
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
playwright==1.38.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pyee==9.0.4
Pygments==2.16.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
selenium==4.13.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tkcalendar==1.6.1
tornado==6.3.3
traitlets==5.10.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
Unidecode==1.3.7
urllib3==2.0.5
wcwidth==0.2.8
webdriver-manager==4.0.1
wsproto==1.2.0
zstandard==0.21.0
